migration: refactor units and configure
When a large database is loaded, the nextcloud-db container takes a long
time to start and causes the failure of nextcloud-app start.

This commit removes strict dependencies between systemd units: units can
now be started selectively during the migration phase.

Also, during migration nextcloud-db is started with database dump and
killed at the end to avoid parallel execution with nextcloud-app.